requirements.txt
absl-py==2.1.0
attrs==23.2.0
cffi==1.16.0
contourpy==1.2.1
cycler==0.12.1
flatbuffers==24.3.25
fonttools==4.51.0
jax==0.4.28
jaxlib==0.4.28
keyboard==0.13.5
kiwisolver==1.4.5
matplotlib==3.9.0
mediapipe==0.10.14
ml-dtypes==0.4.0
MouseInfo==0.1.3
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pillow==10.3.0
protobuf==4.25.3
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python3-xlib==0.15
pytweening==1.2.0
scipy==1.13.0
six==1.16.0
sounddevice==0.4.6
